implemented transfer

+++ app/controllers/api/v1/individuals_controller.rb
@@ -4,7 +4,15 @@ module Api
       before_action :authenticate_api_v1_user!
       include ActionController::MimeResponds
       def index
-        @individuals = Individual.all
+        only_unshipped = params[:unshipped]
+        only_shipped = params[:shipped]
+        if only_unshipped == "true"
+          @individuals = Individual.where(shipped: false)
+        elsif only_shipped == 'true'
+          @individuals = Individual.where(shipped: true)
+        else
+          @individuals = Individual.all
+        end
         @d2=Date.today
 
         respond_to do |format|
@@ -59,19 +67,22 @@ module Api
 
         individual_tags = params[:individual_tags]
 
-                
         @individual_tag_entries = @individual.individual_tag_entries
 
-        @individual_tag_entries.each do |tag_entry|
-          if !individual_tags.include?(tag_entry.individual_tag_id)
-            tag_entry.destroy
+        if @individual_tag_entries.present?
+          @individual_tag_entries.each do |tag_entry|
+            if !individual_tags.include?(tag_entry.individual_tag_id)
+              tag_entry.destroy
+            end
           end
         end
-        
-        individual_tags.each do |tag_id|
-          if @individual_tag_entries.find_by(individual_tag_id: tag_id).nil?
-            @new_individual_tag_entry = @individual.individual_tag_entries.build({individual_tag_id: tag_id})
-            @new_individual_tag_entry.save
+
+        if individual_tags.present? 
+          individual_tags.each do |tag_id|
+            if @individual_tag_entries.find_by(individual_tag_id: tag_id).nil?
+              @new_individual_tag_entry = @individual.individual_tag_entries.build({individual_tag_id: tag_id})
+              @new_individual_tag_entry.save
+            end
           end
         end
